1.[REFACTORING]:move header files to seperate dir 2.[NEW FEATURE IN COMING]:preparing for scene
